Small fix
- Del "warehouse demo" folder
- Solved twice opened problem (in app.py)
- Recreate app.exe
