+++ Create exe.bat
@@ -1,0 +1,4 @@
+pyinstaller --noconfirm --onefile --add-data "schema.sql;." --add-data "data.db;." --add-data "static;static" --hidden-import sqlite3 app.py
+
+
+
