+++ LICENSE
@@ -1,0 +1,6 @@
+MIT License
+
+Copyright (c) 2025
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+... (standard MIT license text truncated for brevity) ...
