+++ README.md
@@ -1,0 +1,25 @@
+"Henry, where my socks?"
+
+Демонстраційний проєкт: Система автоматизації складу (локальна, demo).
+
+Вміст:
+- app.py               - Flask сервер з REST API
+- schema.sql           - SQL-схема бази
+- static/index.html    - простий фронтенд (HTML/JS)
+- static/app.js        - логіка фронтенду
+- static/styles.css    - оформлення сторінки
+- data.db              - база даних (створюється автоматично)
+- requirements.txt     - залежності
+- sample_data.txt      - приклади запитів (curl)
+
+Як запустити (локально):
+1) Переконайтесь, що встановлено Python 3.8+
+2) Створіть віртуальне середовище (рекомендується): python -m venv venv
+3) Активуйте вірт. середовище і встановіть залежності: pip install -r requirements.txt
+4) Запустіть сервер: python app.py
+5) Відкрийте в браузері: http://localhost:5000
+
+Примітки:
+- Проєкт створений виключно для демонстрації.
+- База data.db створюється автоматично при першому запуску.
+- Є мокап даних при ініціалізації.
