+++ requirements.txt
@@ -1,0 +1,1 @@
+Flask>=2.0
