+++ sample_data.txt
@@ -1,0 +1,13 @@
+Приклади curl:
+
+Додати товар:
+curl -X POST -H "Content-Type: application/json" -d '{"sku":"SKU100","name":"Дриль"}' http://localhost:5000/api/products
+
+Додати локацію:
+curl -X POST -H "Content-Type: application/json" -d '{"code":"C1","description":"Полиця C1"}' http://localhost:5000/api/locations
+
+Прийняти партію:
+curl -X POST -H "Content-Type: application/json" -d '{"sku":"SKU001","location":"A1","qty":10,"batch":"BATCH001"}' http://localhost:5000/api/receive
+
+Зняти зі складу:
+curl -X POST -H "Content-Type: application/json" -d '{"sku":"SKU001","qty":3,"location":"A1"}' http://localhost:5000/api/pick
