--- warehouse_demo/LICENSE
@@ -1,6 +1,0 @@
-MIT License
-
-Copyright (c) 2025
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-... (standard MIT license text truncated for brevity) ...
