--- warehouse_demo/requirements.txt
@@ -1,1 +1,0 @@
-Flask>=2.0
