--- warehouse_demo/schema.sql
@@ -1,39 +1,0 @@
-PRAGMA foreign_keys = ON;
-
-CREATE TABLE IF NOT EXISTS products (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    sku TEXT UNIQUE NOT NULL,
-    name TEXT NOT NULL,
-    unit TEXT DEFAULT 'шт'
-);
-
-CREATE TABLE IF NOT EXISTS locations (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    code TEXT UNIQUE NOT NULL,
-    description TEXT
-);
-
-CREATE TABLE IF NOT EXISTS stock (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    product_id INTEGER NOT NULL,
-    location_id INTEGER NOT NULL,
-    batch TEXT,
-    quantity INTEGER NOT NULL DEFAULT 0,
-    UNIQUE(product_id, location_id, batch),
-    FOREIGN KEY(product_id) REFERENCES products(id) ON DELETE CASCADE,
-    FOREIGN KEY(location_id) REFERENCES locations(id) ON DELETE CASCADE
-);
-
-CREATE TABLE IF NOT EXISTS transactions (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    type TEXT NOT NULL,
-    product_id INTEGER NOT NULL,
-    location_id INTEGER,
-    qty INTEGER NOT NULL,
-    batch TEXT,
-    ts DATETIME DEFAULT CURRENT_TIMESTAMP,
-    note TEXT,
-    FOREIGN KEY(product_id) REFERENCES products(id),
-    FOREIGN KEY(location_id) REFERENCES locations(id)
-);
-
