--- warehouse_demo/static/index.html
@@ -1,106 +1,0 @@
-<!DOCTYPE HTML PUBLIC>
-<html lang="uk">
-
-<head>
-  <meta charset="utf-8" />
-  <meta name="viewport" content="width=device-width,initial-scale=1">
-  <title>Demo: Автоматизація складу (локально)</title>
-  <link rel="stylesheet" href="../static/styles.css">
-</head>
-
-<body>
-  <div class="container">
-    <div class="theme-switch">
-      Темна тема
-      <label class="switch" for="theme-toggle">
-        <input type="checkbox" id="theme-toggle">
-        <span class="slider round"></span>
-      </label>
-
-    </div>
-    <div class="container">
-      <h1>Demo: Автоматизація складу</h1>
-      <div class="grid">
-        <div class="card">
-          <h2>Товари</h2>
-          <form id="product-form">
-            <input name="sku" placeholder="SKU" required>
-            <input name="name" placeholder="Назва" required>
-            <input name="unit" placeholder="Од." value="шт">
-            <button type="submit">Додати товар</button>
-          </form>
-          <div id="products"></div>
-        </div>
-        <div class="card">
-          <h2>Зберігання</h2>
-          <form id="location-form">
-            <input name="code" placeholder="Код (A1)" required>
-            <input name="desc" placeholder="Опис">
-            <button type="submit">Додати локацію</button>
-          </form>
-          <div id="locations"></div>
-        </div>
-        <div class="card">
-          <h2>Приймання</h2>
-          <form id="receive-form">
-            <input name="sku" placeholder="SKU" required>
-            <input name="location" placeholder="Локація (A1)" required>
-            <input name="qty" type="number" placeholder="Кількість" required min="1" value="1">
-            <input name="batch" placeholder="Партія (опц.)">
-            <button type="submit">Прийняти</button>
-          </form>
-        </div>
-        <div class="card">
-          <h2>Видача</h2>
-          <form id="pick-form">
-            <input name="sku" placeholder="SKU" required>
-            <input name="location" placeholder="Локація (опц.)">
-            <input name="qty" type="number" placeholder="Кількість" required min="1" value="1">
-            <input name="batch" placeholder="Партія (опц.)">
-            <button type="submit">Віддати</button>
-          </form>
-        </div>
-      </div>
-
-      <div class="card wide">
-        <h2>Залишки на складі</h2>
-        <button class="refresh-btn" id="refresh-stock">Оновити</button>
-        <table id="stock-table">
-          <thead>
-            <tr>
-              <th>SKU</th>
-              <th>Назва</th>
-              <th>Локація</th>
-              <th>Партія</th>
-              <th>Кількість</th>
-            </tr>
-          </thead>
-          <tbody></tbody>
-        </table>
-      </div>
-
-      <div class="card wide">
-        <h2>Транзакції</h2>
-        <button class="refresh-btn" id="refresh-trans">Оновити</button>
-        <div id="transactions"></div>
-      </div>
-    </div>
-    <script>
-      const toggle = document.getElementById('theme-toggle');
-      const body = document.body;
-
-      // Встановлюємо тему при завантаженні сторінки
-      if (localStorage.getItem('darkTheme') === 'true') {
-        body.classList.add('dark');
-        toggle.checked = true;
-      }
-
-      toggle.addEventListener('change', () => {
-        body.classList.toggle('dark');
-        localStorage.setItem('darkTheme', body.classList.contains('dark'));
-      });
-    </script>
-    <script src="../static/app.js"></script>
-</body>
-
-</html>
